+++ scopes/harmony/cli/command-runner.ts
@@ -1,12 +1,10 @@
 import { render } from 'ink';
-import { serializeError } from 'serialize-error';
 import { migrate } from '@teambit/legacy/dist/api/consumer';
 import logger, { LoggerLevel } from '@teambit/legacy/dist/logger/logger';
-import { buildCommandMessage, isNumeric, packCommand } from '@teambit/legacy/dist/utils';
 import { CLIArgs, Command, Flags } from '@teambit/legacy/dist/cli/command';
 import { parseCommandName } from '@teambit/legacy/dist/cli/command-registry';
-import defaultHandleError from '@teambit/legacy/dist/cli/default-error-handler';
 import loader from '@teambit/legacy/dist/cli/loader';
+import { handleErrorAndExit } from '@teambit/legacy/dist/cli/handle-errors';
 
 export class CommandRunner {
   private commandName: string;
@@ -119,44 +117,4 @@ export class CommandRunner {
     }
     return null;
   }
-}
-
-function serializeErrAndExit(err, commandName: string) {
-  const data = packCommand(buildCommandMessage(serializeError(err), undefined, false), false, false);
-  const code = err.code && isNumeric(err.code) ? err.code : 1;
-  // eslint-disable-next-line @typescript-eslint/no-misused-promises
-  return process.stderr.write(data, () => logger.exitAfterFlush(code, commandName));
-}
-
-export async function handleErrorAndExit(err: Error, commandName: string, shouldSerialize = false): Promise<void> {
-  try {
-    loader.off();
-    logger.error(`got an error from command ${commandName}: ${err}`);
-    logger.error(err.stack || '<no error stack was found>');
-    const { message, error } = defaultHandleError(err);
-    if (shouldSerialize) serializeErrAndExit(error, commandName);
-    else await logErrAndExit(message, commandName);
-  } catch (e) {
-    // eslint-disable-next-line no-console
-    console.error('failed to log the error properly, failure error', e);
-    // eslint-disable-next-line no-console
-    console.error('failed to log the error properly, original error', err);
-    process.exit(1);
-  }
-}
-
-export async function handleUnhandledRejection(err: Error | null | undefined | {}) {
-  // eslint-disable-next-line no-console
-  console.error('** unhandled rejection found, please make sure the promise is resolved/rejected correctly! **');
-  if (err instanceof Error) {
-    return handleErrorAndExit(err, process.argv[2]);
-  }
-  console.error(err); // eslint-disable-line
-  return handleErrorAndExit(new Error(`unhandledRejections found. err ${err}`), process.argv[2]);
-}
-
-export async function logErrAndExit(err: Error | string, commandName: string) {
-  if (!err) throw new Error(`logErrAndExit expects to get either an Error or a string, got nothing`);
-  console.error(err); // eslint-disable-line
-  await logger.exitAfterFlush(1, commandName);
 }
